+++ CHANGES.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## 1.3.0 (2016-??-??)
+* Add support for `Disc#toc_string`. Will use the implementation from libdiscid
+  0.6 or a fallback for older versions.
+
 ## 1.2.0 (2016-02-12)
 * `DiscId.sectors_to_seconds`: Truncate instead of round resulting float